Fix : 사장 admin -> owner 수정
