+++ src/main/java/org/example/fourchak/domain/auth/dto/request/SignupRequest.java
@@ -1,5 +1,4 @@
-package org.example.fourchak.auth.dto.request;
-
+package org.example.fourchak.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -11,15 +10,23 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class SigninRequest {
+public class SignupRequest {
 
     @Email
     @NotBlank
     private String email;
 
     @NotBlank
+    private String username;
+
+    @NotBlank
+    private String phone;
+
+    @NotBlank
     @Pattern(regexp = "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)[A-Za-z\\d]{8,}$",
         message = "비밀번호는 소영문,대영문자와 숫자를 포함한 8자 이상이어야 합니다.")
     private String password;
 
+    @NotBlank
+    private String userRole;
 }
